modify some logic for alias save and init
