+++ README.zh-CN.md
@@ -14,7 +14,7 @@ Go通用的数据验证与过滤库，使用简单，内置大部分常用验证
 - 简单方便，支持前置验证检查, 支持添加自定义验证器
 - 支持将规则按场景进行分组设置，不同场景验证不同的字段
 - 支持在进行验证前对值使用过滤器进行净化过滤，查看 [内置过滤器](#built-in-filters)
-- 已经内置了超多（> 60 个）常用的验证器，查看 [内置验证器](#built-in-validators)
+- 已经内置了超多（**>70** 个）常用的验证器，查看 [内置验证器](#built-in-validators)
 - 方便的获取错误信息，验证后的安全数据获取(只会收集有规则检查过的数据)
 - 支持自定义每个验证的错误消息，字段翻译，消息翻译(内置`en` `zh-CN`)
 - 完善的单元测试，测试覆盖率 > 90%
